+++ packages/components-library/src/components/ui/Pagination.vue
@@ -4,13 +4,13 @@
   <div v-if="localValue.count > localValue.size" class="btn-group mb-2 mr-3">
     <button :class="css" class="btn btn-outline-primary"
       :disabled="localValue.page <= 1"
-      @click="navigate(1)">
+      @click="updateValue({page: 1})">
       «
     </button>
 
     <button :class="css" class="t-page-prev btn btn-outline-primary"
       :disabled="localValue.page <= 1"
-      @click="navigate(localValue.page - 1)">
+      @click="updateValue({page: localValue.page - 1})">
       ‹
     </button>
 
@@ -18,21 +18,21 @@
       v-for="pageNumber in pageNumbers" :key="pageNumber"
       class="btn btn-outline-primary"
       :class="{'active': pageNumber === localValue.page, ...css}"
-      @click="navigate(pageNumber)">
+      @click="updateValue({page: pageNumber})">
       {{pageNumber}}
     </button>
 
     <button
       :class="css" class="t-page-next btn btn-outline-primary"
       :disabled="localValue.page >= pageCount"
-      @click="navigate(localValue.page + 1)">
+      @click="updateValue({page: localValue.page + 1})">
       ›
     </button>
 
     <button
       :class="css" class="btn btn-outline-primary"
       :disabled="localValue.page >= pageCount"
-      @click="navigate(pageCount)">
+      @click="updateValue({page: pageCount})">
       »
     </button>
   </div>
@@ -44,7 +44,7 @@
 
     <div class="form-check form-check-inline">
       <label class="form-check-label mr-2" for="page-size">{{i18n['per page']}}</label>
-      <select class="form-control" id="page-size" v-model="localValue.size">
+      <select class="form-control" id="page-size" v-model="localValue.size" @change="handleSizeChange">
         <option v-for="pageSize of pageSizes" :value="pageSize" :key="pageSize">{{pageSize}}</option>
       </select>
     </div>
@@ -55,16 +55,62 @@
 
 <script>
 export default {
-  async created () {
-    this.localValue = { ...this.localValue, ...this.value }
-
-    // Life-cycle hook for router-less pagination.
-    if (!this.useRouter) {
-      // Allow Vue instance listeners to initialize in tests.
-
-      // Pagination state is leading for the initial page,
-      // in case no router is being use.
-      this.fetchData(this.localValue.page)
+  name: 'Pagination',
+  props: {
+    /**
+     * Extra css classes for the pagination buttons.
+     */
+    css: {
+      type: Object,
+      required: false,
+      default: () => ({})
+    },
+    /**
+     * Translatable texts.
+     */
+    i18n: {
+      type: Object,
+      default: () => ({
+        page: 'page',
+        'per page': 'per page',
+        items: 'items'
+      })
+    },
+    /**
+     * Select choices for the page size.
+     */
+    pageSizes: {
+      type: Array,
+      required: false,
+      default: () => [10, 20, 50]
+    },
+    /**
+     * Reflects the pagination state.
+     * @model
+     */
+    value: {
+      type: Object,
+      required: true
+    },
+    /**
+     * Number of navigational page buttons; this must be uneven.
+     */
+    visiblePages: {
+      type: Number,
+      default: () => 7,
+      validator (value) {
+        return (value % 2 !== 0)
+      }
+    }
+  },
+  data: function () {
+    return {
+      localValue: {
+        size: 20,
+        page: 1,
+        loading: false,
+        count: 0
+      }
     }
   },
   computed: {
@@ -110,153 +156,28 @@ export default {
       return { left, right }
     }
   },
-  data: function () {
-    return {
-      localValue: { size: 20, page: 1, loading: false, count: 0 }
-    }
-  },
   methods: {
-    async fetchData (page) {
-      // Omit the Pagination method when you're using multiple
-      // Pagination components with the same data source.
-      if (!this.fetchItems) { return }
-
-      this.updateValue({ loading: true, page })
-      await this.fetchItems()
-      this.updateValue({ loading: false })
-    },
-    navigate (page) {
-      if (!this.useRouter) { return this.fetchData(page) }
-
-      if (this.$route.query.page) {
-        if (
-          parseInt(this.$route.query.page, 10) !== page ||
-          parseInt(this.$route.query.size, 10) !== this.value.size
-        ) {
-          // (!) Merge with existing query; e.g. bookmarks
-          this.$router.push({
-            path: this.$route.path,
-            query: { ...this.$route.query, page, size: this.value.size }
-          })
-        }
-      } else {
-        this.$router.replace({
-          path: this.$route.path,
-          query: { ...this.$route.query, page, size: this.value.size }
-        })
-      }
-    },
     updateValue (value) {
       this.localValue = { ...this.localValue, ...this.value, ...value }
-      this.$emit('input', { ...this.value, ...this.localValue, ...value })
-    }
-  },
-  name: 'Pagination',
-  props: {
-    /**
-     * Extra css classes for the pagination buttons.
-     */
-    css: {
-      type: Object,
-      required: false,
-      default: () => ({})
+      this.$emit('input', { ...this.localValue })
     },
-    /**
-     * Async method to retrieve items with.
-     * @returns {Object} {count: Number}
-     */
-    fetchItems: {
-      type: Function,
-      required: false
-    },
-    /**
-     * Translatable texts.
-     */
-    i18n: {
-      type: Object,
-      default: () => ({
-        page: 'page',
-        'per page': 'per page',
-        items: 'items'
-      })
-    },
-    /**
-     * Select choices for the page size.
-     */
-    pageSizes: {
-      type: Array,
-      required: false,
-      default: () => [10, 20, 50]
-    },
-    /**
-     * Optional vue-router integration.
-     */
-    useRouter: {
-      type: Boolean,
-      default: () => true
-    },
-    /**
-     * Reflects the pagination state.
-     * @model
-     */
-    value: {
-      type: Object,
-      required: true
-    },
-    /**
-     * Number of navigational page buttons; this must be uneven.
-     */
-    visiblePages: {
-      type: Number,
-      default: () => 7,
-      validator (value) {
-        return (value % 2 !== 0)
-      }
+    handleSizeChange () {
+      this.updateValue({ page: 1, size: this.localValue.size })
     }
   },
   watch: {
-    '$route.query': {
-      handler: function (query) {
-        if (!this.fetchItems || !this.useRouter) { return }
-
-        if (query.page) {
-          this.fetchData(parseInt(query.page, 10))
-        } else {
-          // Defaults by setting the pagination to the first page.
-          this.navigate(1)
-        }
-      },
-      deep: true,
-      immediate: true
-    },
     /**
      * Sync the local state when the external state changes.
      */
     value: {
-      handler (value) { this.localValue = value },
+      handler (value) {
+        this.localValue = value
+      },
       deep: true
-    },
-    /**
-     * A separate listener is needed here, because its
-     * not feasible to detect size property changes from
-     * the deep value watcher (See https://vuejs.org/v2/api/#vm-watch)
-     */
-    'value.size' (newSize, oldSize) {
-      // Only navigate when the paginator is not in passive mode.
-      if (this.fetchItems) {
-        let currentPosition = this.localValue.page * oldSize
-        // Keep in mind that with 201 items, the current page may be 3 when
-        // showing 100 items per page. When switching back to 50 items per
-        // page, there are only 5 pages left. Make sure the currentPosition
-        // never exceeds the actual count.
-        if (currentPosition > this.localValue.count) {
-          currentPosition = this.localValue.count
-        }
-
-        const currentNewPage = Math.ceil(currentPosition / newSize)
-        this.navigate(currentNewPage)
-      }
     }
+  },
+  async created () {
+    this.localValue = { ...this.localValue, ...this.value }
   }
 }
 </script>
@@ -302,22 +223,14 @@ export default {
 ```jsx
   // Always provide at least 'size' and 'page'; the other properties
   // are added from the pagination component itself.
-  let model = { size: 20, page: 1 }
+  let model = { size: 20, page: 1, count: 124 }
   const mockStore = (e) => { model = e }
-
-  const fetchItems = async function() {
-    // Retrieve items and feed the pagination component
-    // with the api results for paginated content.
-    return {count: 250 }
-  }
 
   <Pagination
     v-model="model"
-    v-bind:fetchItems="fetchItems"
     v-on:input="mockStore"
     v-bind:visiblePages="9"
     v-bind:pageSizes="[10, 20, 50, 100]"
-    v-bind:useRouter="false"
     v-bind:css="{'btn-sm': true}"/>
   Model: {{model}}
 ```