+++ packages/data-explorer/src/types/Sort.ts
@@ -1,0 +1,4 @@
+export type Sort = {
+  sortColumnName: string | null,
+  isSortOrderReversed: boolean,
+}