+++ packages/data-explorer/src/views/MainView.vue
@@ -40,7 +40,6 @@
         <pagination
           class="mt-2"
           v-model="tablePagination"
-          :fetchItems="() => fetchViewData({ tableName: $route.params.entity })"
         />
         </div>
     </div>
@@ -73,7 +72,13 @@ export default Vue.extend({
       get: function () {
         return this.$store.state.tablePagination
       },
-      set: function (value) { this.setPagination(value) }
+      set: function (value) {
+        this.$router.push({
+          name: this.$router.currentRoute.name,
+          path: this.$router.currentRoute.path,
+          query: { ...this.$route.query, page: value.page, size: value.size }
+        })
+      }
     },
     toasts: {
       get: function () {
@@ -98,7 +103,7 @@ export default Vue.extend({
     ]),
     ...mapGetters([
       'isUserAuthenticated',
-      'compressedBookmark'
+      'compressedRouteFilter'
     ]),
     activeFilterSelections () {
       return this.searchText ? { ...this.filters.selections, _search: this.searchText } : this.filters.selections
@@ -117,14 +122,14 @@ export default Vue.extend({
       'setHideFilters',
       'setTableName',
       'setToasts',
-      'setPagination',
       'setSearchText',
       'setFilterSelection',
-      'applyBookmark'
+      'setRouteQuery'
     ]),
     ...mapActions([
       'deleteRow',
-      'fetchViewData'
+      'fetchViewData',
+      'fetchTableMeta'
     ]),
     ...mapActions('header', [
       'fetchPackageTables'
@@ -137,7 +142,7 @@ export default Vue.extend({
       this.$router.push({
         name: this.$router.currentRoute.name,
         path: this.$router.currentRoute.path,
-        query: { bookmark: this.compressedBookmark }
+        query: { ...this.$route.query, filter: this.compressedRouteFilter }
       })
     },
     async handeldeleteItem (itemId) {
@@ -148,26 +153,24 @@ export default Vue.extend({
       }
     }
   },
-  created () {
+  async created () {
     this.$eventBus.$on('delete-item', (data) => {
       this.handeldeleteItem(data)
     })
+    await this.fetchTableMeta({ tableName: this.$route.params.entity })
+    this.setRouteQuery(this.$route.query)
+    this.fetchViewData()
   },
   destroyed () {
     this.$eventBus.$off('delete-item')
   },
   async beforeRouteUpdate (to, from, next) {
-    if (this.$route.params.entity !== to.params.entity) {
-      // Reset pagination to defaults before loading another entity.
-      this.setPagination()
-      await this.fetchViewData({ tableName: to.params.entity })
+    if (to.params.entity !== from.params.entity) {
+      await this.fetchTableMeta({ tableName: to.params.entity })
     }
+    this.setRouteQuery(to.query) // syncs the state with the query
+    await this.fetchViewData()
     next()
-  },
-  watch: {
-    '$route.query': function (query) {
-      this.applyBookmark(query.bookmark || '')
-    }
   }
 })
 </script>